refactor: remove duplicate code and clean up backend
- Remove node-cron dependency from backend package.json
- Replace old SyncService with new integrated version
- Replace old sync routes with new integrated version
- Remove duplicate scheduler/cron code
- Maintain backward compatibility
- Clean up imports and route references
