+++ backend/package.json
@@ -55,7 +55,6 @@
     "ioredis": "^5.3.2",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
-    "node-cron": "^3.0.3",
     "pino": "^8.17.2",
     "pino-pretty": "^10.2.3",
     "prisma": "^5.7.1",
@@ -75,7 +74,6 @@
     "@types/jest": "^29.5.8",
     "@types/jsonwebtoken": "^9.0.5",
     "@types/lodash": "^4.14.202",
-    "@types/node-cron": "^3.0.11",
     "@types/supertest": "^2.0.16",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui-express": "^4.1.6",
