+++ backend/src/index.ts
@@ -23,7 +23,6 @@ import stackRoutes from '@/routes/stacks';
 import developerRoutes from '@/routes/developers';
 import repositoryRoutes from '@/routes/repositories';
 import syncRoutes from '@/routes/sync';
-import syncNewRoutes from '@/routes/syncNew';
 import schedulerRoutes from '@/routes/scheduler';
 import kpiRoutes from '@/routes/kpis';
 import azureDevOpsRoutes from '@/routes/azureDevOps';
@@ -159,7 +158,6 @@ app.use('/api/stacks', stackRoutes);
 app.use('/api/developers', developerRoutes);
 app.use('/api/repositories', repositoryRoutes);
 app.use('/api/sync', syncRoutes);
-app.use('/api/sync-new', syncNewRoutes);
 app.use('/api/scheduler', schedulerRoutes);
 app.use('/api/kpis', kpiRoutes);
 app.use('/api/azure-devops', azureDevOpsRoutes);
@@ -182,7 +180,6 @@ app.get('/api', (req, res) => {
       developers: '/api/developers',
       repositories: '/api/repositories',
       sync: '/api/sync',
-      'sync-new': '/api/sync-new',
       scheduler: '/api/scheduler',
       kpis: '/api/kpis',
     },
